+++ CHANGELOG.md
@@ -21,6 +21,7 @@
 - Fixed a recursion error on duplicate dependencies with only different extras.
 - Fixed handling of extras.
 - Fixed duplicate entries in both sdist and wheel.
+- Fixed excluded files appearing in the `package_data` of the generated `setup.py`.
 
 
 ## [0.11.5] - 2018-09-04